requirements.txt
-e .
cdk-workshop-coffee-listing-app==0.1.0